libs/python/odg-core/pyproject.toml
[project]
name = "odg-core"
version = "0.1.0"
description = "Shared models, enums, and utilities for OpenDataGov"
requires-python = ">=3.13"
license = "Apache-2.0"
dependencies = [
    "pydantic>=2.0,<3.0",
    "sqlalchemy[asyncio]>=2.0,<3.0",
    "asyncpg>=0.29,<1.0",
    "redis>=5.0,<8.0",
    "alembic>=1.13,<2.0",
    "pydantic-settings>=2.0,<3.0",
    "opentelemetry-api>=1.24,<2.0",
    "opentelemetry-sdk>=1.24,<2.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24,<2.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.45b0",
    "opentelemetry-instrumentation-httpx>=0.45b0",
    "opentelemetry-instrumentation-redis>=0.45b0",
    "pyyaml>=6.0,<7.0",
    "aiokafka>=0.10,<1.0",
    "hvac>=2.1,<3.0",
    "PyJWT[crypto]>=2.8,<3.0",
    "openlineage-python>=1.12,<2.0",
    "acryl-datahub>=0.12,<1.0",
    "opendp>=0.10,<1.0",
    "strawberry-graphql[fastapi]>=0.200,<1.0",
    "nats-py>=2.0,<3.0",
    "grpcio>=1.60,<2.0",
    "grpcio-tools>=1.60,<2.0",
    "protobuf>=4.25,<6.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=1.0",
    "pytest-cov>=7.0",
    "ruff>=0.15",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
    "pyarrow-stubs",
    "types-beautifulsoup4",
    "duckdb",
    "feast",
    "kafka-python",
    "couchdb",
    "boto3-stubs",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/odg_core"]

[tool.ruff]
target-version = "py313"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "A", "SIM", "TCH", "RUF"]

[tool.ruff.lint.per-file-ignores]
"**/graphql/**/*.py" = ["TCH"]      # Strawberry evaluates annotations at runtime
"**/graphql/schema.py" = ["A002", "TCH"]  # GraphQL resolvers use `id` parameter; strawberry runtime annotations
"**/api/*.py" = ["TCH", "B008"]    # FastAPI evaluates annotations at runtime
"**/grpc/*.py" = ["N802"]          # gRPC methods follow PascalCase convention

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["hvac", "hvac.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["opendp", "opendp.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["numpy", "numpy.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["aiokafka", "aiokafka.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["great_expectations", "great_expectations.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["datahub", "datahub.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["openlineage", "openlineage.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["fastapi", "fastapi.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["scipy", "scipy.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["strawberry", "strawberry.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["nats", "nats.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["grpc", "grpc.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["google", "google.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["kafka", "kafka.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["couchdb", "couchdb.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["gremlin_python", "gremlin_python.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["bs4", "bs4.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["minio", "minio.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["pyiceberg", "pyiceberg.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["pandas", "pandas.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["requests", "requests.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["httpx", "httpx.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["feast", "feast.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["pyspark", "pyspark.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["odg_core.proto_gen.*"]
ignore_missing_imports = true
ignore_errors = true

[[tool.mypy.overrides]]
module = ["airflow", "airflow.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["polars", "polars.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["duckdb", "duckdb.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["pyarrow", "pyarrow.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["boto3", "boto3.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
warn_return_any = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
